example update: replaced with typescript
